+++ pages/users.py
@@ -1,9 +1,20 @@
 import pandas as pd
-import plotly.express as px
 
 import dash
-from dash import dcc, Output, Input, html, dash_table, State, ctx
-from src.scripts import *
+from dash import dcc, Output, Input, html, State, ctx
+from src.data_functions import (
+    create_fy_options,
+    merge_workbooks,
+    get_date_list,
+    select_df,
+    get_totals,
+)
+from src.ui_functions import (
+    make_df_download_button,
+    make_summary_panel,
+    make_data_table,
+    make_bar_graph,
+)
 import logging
 
 from config import settings
@@ -25,54 +36,19 @@ WORKSHEETS = [
 
 DATAFRAMES = merge_workbooks(WORKSHEETS)
 
-download_button = html.Div(
-    children=[
-        html.Button(
-            "Download Data",
-            id="btn-download",
-            className="c-button c-button--primary btn-download",
-        ),
-        html.Hr(),
-        dcc.Download(id="download-users-df"),
-    ],
-)
 
 # CUSTOMIZE LAYOUT
 layout = html.Div(
     [
         html.Div(
             [
-                html.Div(
-                    [
-                        html.Div(
-                            [
-                                html.Div(
-                                    ["Average Total Users"], className="counter_title"
-                                ),
-                                html.Div([0], id="total_users"),
-                            ],
-                            className="total_counters",
-                        ),
-                        html.Div(
-                            [
-                                html.Div(["Average Active"], className="counter_title"),
-                                html.Div([0], id="active_users"),
-                            ],
-                            className="total_counters",
-                        ),
-                        html.Div(
-                            [
-                                html.Div(["Average Idle"], className="counter_title"),
-                                html.Div([0], id="idle_users"),
-                            ],
-                            className="total_counters",
-                        ),
-                    ],
-                    id="total_counters_wrapper",
+                make_summary_panel(
+                    ["Average Total Users", "Average Active", "Average Idle"],
+                    ["total_users", "active_users", "idle_users"],
                 ),
                 html.Div(children=[], id="bargraph"),
                 html.Div(children=[], id="table"),
-                download_button,
+                make_df_download_button("users"),
             ],
         ),
         dcc.Location(id="url"),
@@ -80,6 +56,7 @@ layout = html.Div(
 )
 
 
+# ADD INTERACTIVITY THROUGH CALLBACKS
 @app.callback(
     Output("download-users-df", "data"),
     Input("btn-download", "n_clicks"),
@@ -110,7 +87,6 @@ def func(
     return dcc.send_data_frame(df.to_csv, "utrc_data.csv")
 
 
-# ADD INTERACTIVITY THROUGH CALLBACKS
 @app.callback(
     Output("table", "children"),
     Output("bargraph", "children"),
@@ -130,6 +106,7 @@ def update_figs(
     start_date,
     end_date,
 ):
+    logging.debug(f"Callback trigger id: {ctx.triggered_id}")
     dates = get_date_list(start_date, end_date)
     df = select_df(
         DATAFRAMES,
@@ -139,23 +116,7 @@ def update_figs(
         machines,
     )
 
-    styles = get_table_styles()
-
-    table = dash_table.DataTable(
-        id="datatable_id",
-        data=df.to_dict("records"),
-        columns=[{"name": i, "id": i} for i in df.columns],
-        fixed_rows={"headers": True},
-        page_size=200,
-        style_header=styles["style_header"],
-        style_cell=styles["style_cell"],
-        style_data_conditional=styles["style_data_conditional"],
-        style_cell_conditional=create_conditional_style(df),
-        style_header_conditional=styles["style_header_conditional"],
-        sort_action="native",
-        filter_action="native",
-        style_as_list_view=True,
-    )
+    table = make_data_table(df)
 
     inst_grps = df.groupby(["Institution"])
     df_with_avgs = {"Institution": [], "Date": []}
@@ -169,42 +130,8 @@ def update_figs(
             continue  # For some date ranges, even if an institution is checked, it doesn't appear in the data, throwing an error
     combined_df = pd.concat([df, pd.DataFrame(df_with_avgs)])
 
-    colors = make_color_map(dates)
-    print(colors)
-
-    bargraph = html.Div(
-        [
-            html.H2("Users per Institution"),
-            dcc.Graph(
-                figure=px.histogram(
-                    data_frame=combined_df,
-                    x="Institution",
-                    color="Date",
-                    barmode="group",
-                    color_discrete_map=colors,
-                    text_auto=True,
-                    category_orders={
-                        "Institution": [
-                            "UTAus",
-                            "UTA",
-                            "UTD",
-                            "UTEP",
-                            "UTPB",
-                            "UTRGV",
-                            "UTSA",
-                            "UTT",
-                            "UTHSC-H",
-                            "UTHSC-SA",
-                            "UTMB",
-                            "UTMDA",
-                            "UTSW",
-                            "UTSYS",
-                        ]
-                    },
-                ).update_layout(yaxis_title="Number of Users")
-            ),
-        ],
-        className="graph-card",
+    bargraph = make_bar_graph(
+        combined_df, "Users per Institution", dates, None, "Number of Users"
     )
 
     totals = get_totals(
